v0/requirements.txt
sliver-py
grpcio 
grpcio-tools